Relocate stateless decode helpers to paz.backend.standard
snapshot_variables and call_stateless (was stateless_forward) move to
paz.backend.standard and are re-exported as paz.snapshot_variables /
paz.call_stateless. They are generic "run a Keras model statelessly under jit
with weights passed as arguments" plumbing, not Gemma4-specific. Gemma4's
multimodal_decoding imports them from paz instead of defining local copies.
Behavior unchanged (gemma4 jitted decode / sampling tests identical).
